:recycle: Clean up layout and rename to default-layout to match vue conventions.
