+++ web/src/components/default-layout/LeftSidebarNavigationDrawer.vue
